requirements.txt
Django==5.2.5
asgiref==3.9.1
autopep8==1.6.0
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
django-environ==0.8.1
libsass==0.22.0
psycopg2-binary==2.9.5
django-tables2==2.7.5
requests==2.32.5
